backend/requirements.txt
absl-py==2.3.1
asgiref==3.9.1
astunparse==1.6.3
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cryptography==45.0.6
cycler==0.12.1
dataclasses==0.6
defusedxml==0.7.1
dj-rest-auth==2.2.6
Django==3.2.25
django-allauth==0.49.0
django-cors-headers==3.13.0
django-rest-framework-mongoengine @ git+https://github.com/umutbozkurt/django-rest-framework-mongoengine.git@2e1b2d3ce3ebd54b325b4999fe6efaf75c0aec23
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djongo==1.3.6
filelock==3.19.1
flatbuffers==25.2.10
fonttools==4.59.1
fsspec==2025.7.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.74.0
h5py==3.14.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
keras==3.11.2
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mdurl==0.1.2
ml_dtypes==0.5.3
mongoengine==0.29.1
mpmath==1.3.0
namex==0.1.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
oauthlib==3.3.1
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
protobuf==6.32.0
psutil==7.0.0
py-cpuinfo==9.0.0
pycparser==2.22
Pygments==2.19.2
PyJWT==2.10.1
pymongo==3.12.3
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.1.0
scipy==1.15.3
six==1.17.0
sqlparse==0.2.4
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.179
ultralytics-thop==2.0.15
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.17.3
